LayoutTests/fast/dom/Element/id-in-map.html

<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css">
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>

<map title="map" id="firstmap">

<script>
description("Test using id in map element.");
    
shouldBeNonNull('document.getElementById("firstmap")');

var successfullyParsed = true;
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
